Don't make promises you can't keep
Using an Overlapper for a Floor or Ceil query suggests we can
meaningfully use the Max point of the Overlapper - we cannot since it is
not stored in sort order. Even as it stands we have problems iterating
over a tree using these methods without the user providing additional
discriminators to the Compare method to avoid either repeatedly finding
a target or missing targets with the same start point.