Fast mode for insert and delete
Since child ranges are not used in insertion or deletion there is no
need to update these unless a search operation is used. So to speed up
batch inserts or deletes this range update can be avoided. This means
that we need to provide a tree-wide range update method to clean up
after batch operations.