Short circuit on deletion is not productive
It seems the cost of checking for overlap with a range is greater than
following the branch in the benchmarks as they exist. Note that all
deletions in the benchmarks are guaranteed to be fruitful and the short
circuit is intended to help avoid wasting time on non-fruitful
deletions, however since intervals must have a matching ID to be
deleted, it makes sense to avoid any search by keeping a list of
inserted intervals in another store (hash or slice) if this is an issue.