Use type to determine bounding behaviour
This allows Bounds and Extend to be optional methods for cases where
bounding is required - purely to reduce the client burden when
implementing the interface with new types.

The provided point type privides these so there is still a boolean to
turn off bounding when using this type when bounds are less important
than performance. A non-bounding type is provided in testing.