Don't be over zealous in rejecting valid ranges
Previously we made a zero length range result in a no-op for the generic
case, this stops DoRange finding points that coincide with the range.

This change changes the range from [from, to) to [from, to]. Here this
makes sense. A simple check in the provided Operation can revert the
behaviour.