Correctly calculate child range values
Previously we miscalculated child ranges when a left child's range
extended beyond the end of the range of the right child, resulting in a
child range that may exclude regions actually covered by intervals
stored in the tree. In the worst case, this would result in failure to
recover intervals intersecting with some queries.