Simplify NearestN into special case of NearestSet
Under Go1.0.3 this has a 5% performance cost, but this pretty much goes
away with Go1.1, so for the sake of simplicity we remove NearestN and
publicise the NKeeper type (formerly cDists) to provide the NearestN
functionality.

The performance issues noted in the previous commit were do to
non-pointer receiver use in the now NKeeper type. Indirection is a win
here.