Equality is unnecessary and wasteful
The resulting found point must be less than the current best, and no
point at or on the far side of the hyperplane can be closer than the
current best, so finish here.

The case with NearestSet is more complicated, since the Keeper can - as
is the case with DistKeeper - keep coequal points. So we need to be
permissive there and allow the Keeper to make the decision.